Test refactoring in progress
