+++ tests/README.md
@@ -1,0 +1,3 @@
+Integration tests go here.
+I.e. tests that actually connect to an LDAP server.
+Offline tests should be kept in the source files.
